% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{aibs_dims}
\alias{aibs_dims}
\title{Dimensions for AIBS CCF data at different resolutions}
\usage{
aibs_dims(res = "25um")
}
\arguments{
\item{res}{The resolution of the dataset. Can be "1um","25um", or "200um". Default is "25um".}
}
\value{
a 3-element integer vector with the x,y,z dimensions.

x is anterior<->posterior
y is ventral<->dorsal
z is left<->right
}
\description{
Data at 1um: CCF Meshes
Data at 25um: Annotation grids and most gridded TissueCyte datasets
Data at 200um: Gridded ISH datasets and related annotations
}
