% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{aibs_organism_id}
\alias{aibs_organism_id}
\title{Organism ID used by the AIBS API}
\usage{
aibs_organism_id(species)
}
\arguments{
\item{species}{character, which species to use. Currently available: "human" and "mouse".}
}
\value{
a numeric value used by the AIBS API to represent the selected organism.
}
\description{
Organism ID used by the AIBS API
}
