% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_ccf_meshes.R
\name{ccf_2017_color}
\alias{ccf_2017_color}
\title{Retrieve a set of CCF 2017 colors}
\usage{
ccf_2017_color(acronym = NULL, structure_id = NULL)
}
\arguments{
\item{acronym}{The structure acronym(s) to retrieve. Default is NULL.}

\item{structure_id}{The structure_id(s) to retrieve. Defualt is NULL.}
}
\value{
a vector of color hex values prefixed with "#".
}
\description{
Must provide either acronym(s) or a structure_id(s).
}
\examples{
BLA_mesh <- ccf_2017_color(acronym = "BLA")
}
