% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{cleanup_taxons}
\alias{cleanup_taxons}
\title{Clean up taxons by removing structures no longer present after filtering}
\usage{
cleanup_taxons(flat_ontology, id_column = "id",
  taxon_column = "taxons")
}
\arguments{
\item{flat_ontology}{A flat ontology data.frame generated by flatten_mba_ontology.}

\item{id_column}{The id column used to link structures. default is "id".}

\item{taxon_column}{The name of the column with taxons generated by generate_taxons().}
}
\value{
a data.frame with a modified taxons column that lacks any ids that are no
longer present in the ontology.
}
\description{
Clean up taxons by removing structures no longer present after filtering
}
