% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{compute_hierarchical}
\alias{compute_hierarchical}
\title{Compute values hierarchicaly over a flat ontology}
\usage{
compute_hierarchical(df, fun, compute_column, result_column,
  include_node = FALSE, na.rm = NULL, taxon_column = "taxons")
}
\arguments{
\item{df}{A flat ontology data.frame with a taxons column}

\item{fun}{The function to run as a character object, e.g. "mean".}

\item{compute_column}{The column of values to use for the function computation.}

\item{result_column}{The name of the output column.}

\item{include_node}{Logical, should values assigned to non-leaf nodes be included in calculation?}

\item{na.rm}{If TRUE or FALSE, passed as a parameter to fun. If NULL (default), is ignored.}

\item{taxon_column}{The name of the taxons column. Default is "taxons"}
}
\value{
A flat ontology data.frame with the result_column added. The taxon column will be moved
to the last column position. This should behave similarly to mutate().
}
\description{
Values accumulate from the bottom of the ontology to the top at each node,
and are computed only from the values of the direct children of each node.
}
\examples{

flat_ontology2 <- flat_ontology \%>\%
  compute_hierarchical("mean","n_children","mean_children",
                       include_node = TRUE)

}
