% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{filter_mba_ontology_children}
\alias{filter_mba_ontology_children}
\title{Filter the MBA ontology to find children of a given structure acronym}
\usage{
filter_mba_ontology_children(flat_ontology, parent_acronym,
  include_parent = FALSE)
}
\arguments{
\item{flat_ontology}{A flat ontology data.frame generated by flatten_mba_ontology}

\item{parent_acronym}{The acronym of a brain ontology structure to find}

\item{include_parent}{Whether or not to include the parent structure in the output. Default = FALSE.}
}
\value{
A data.frame with the filtered ontology only for children of the parent structure.
}
\description{
Filter the MBA ontology to find children of a given structure acronym
}
