% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{flatten_mba_ontology}
\alias{flatten_mba_ontology}
\title{Convert a nested Mouse Brain Atlas ontology to a data.frame}
\usage{
flatten_mba_ontology(ontology, ontology_df = NULL)
}
\arguments{
\item{ontology}{A nested ontology object}

\item{ongology_df}{An existing ontology data.frame. This is used for recursion. You should use the default, NULL, to extract a full ontology.}
}
\value{
a data.frame with all descriptive columns for the ontology.
}
\description{
Convert a nested Mouse Brain Atlas ontology to a data.frame
}
