% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{generate_mba_taxons}
\alias{generate_mba_taxons}
\title{Generate a taxon column that shows the "lineage" of each ontology term}
\usage{
generate_mba_taxons(flat_ontology)
}
\arguments{
\item{flat_ontology}{A flat ontology object generated by flatten_mba_ontology}
}
\value{
a data.frame with all descriptive columns for the ontology plus a new taxon column.
}
\description{
Generate a taxon column that shows the "lineage" of each ontology term
}
