% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{generate_taxons}
\alias{generate_taxons}
\title{Generate a taxons column for a flattened ontology}
\usage{
generate_taxons(flat_ontology, id_column = "id",
  parent_column = "parent_structure_id", value_column = "id",
  taxon_column = "taxons")
}
\arguments{
\item{flat_ontology}{A flat ontology data.frame generated by flatten_mba_ontology}

\item{id_column}{The id column used to link structures. default is "id".}

\item{parent_column}{The column specifying the id column of the parent of each structure. default is "parent_structure_id".}

\item{value_column}{The column of values to use to build the taxons column. Default is "id". Using "acronym" will yield a more readable taxonomy.}

\item{taxon_column}{The name of the column to use for storing the resulting taxons. Default is "taxons".}
}
\value{
a data.frame with an additional column of semicolon-separated taxons based on the value_column.
}
\description{
This column is needed for use with the taxa and metacodeR packages.
}
