% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_ish.R
\name{get_aba_ish_data}
\alias{get_aba_ish_data}
\title{Retrieve gridded ABA RNA ISH results for a specific SectionDataSet}
\usage{
get_aba_ish_data(api_id, values = "energy")
}
\arguments{
\item{api_id}{The ID of the SectionDataSet to retrieve}

\item{values}{Which values to retrieve. Options are "energy", "density", "intensity", and "injection". Default = "energy".}
}
\value{
a 3d array with values for each position in the ABA mouse gridAnnotation space at 200 um resolution. Dimension are:
\itemize{
  \item x, 67: anterior/posterior
  \item y, 41: superior/inferior
  \item z, 58: left/right
}
}
\description{
Retrieve gridded ABA RNA ISH results for a specific SectionDataSet
}
