% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_ish.R
\name{get_aba_ish_structure_data}
\alias{get_aba_ish_structure_data}
\title{Get per-structure ABA RNA ISH results for a specific SectionDataSet}
\usage{
get_aba_ish_structure_data(api_id, keep_non_atlas = FALSE)
}
\arguments{
\item{api_id}{The ID of the SectionDataSet to retrieve}

\item{keep_non_atlas}{Logical, whether to keep or filter structures without an atlas id value. Default = FALSE.}
}
\value{
a data.frame with per-structure atlas expression values:
\itemize{
\item name: full name of the atlas structure
\item acronym: structure acronym
\item atlas_id: atlas id value for structure
\item density: normalized expression density value
\item energy: normalized expression as intensity/density
\item sum_expressing_pixel_intensity
\item sum_expressing_pixels
\item sum_pixel_intensity
\item sum_pixels
\item voxel_energy_mean
\item voxel_energy_cv
}
}
\description{
Get per-structure ABA RNA ISH results for a specific SectionDataSet
}
