% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_ccf.R
\name{get_ccf_annotation}
\alias{get_ccf_annotation}
\title{Download the CCF Annotation object and return a 3D Array of values}
\usage{
get_ccf_annotation()
}
\value{
a 3-dimensional array with 25 micron resolution. Dimension are:
\itemize{
  \item x, 528: anterior/posterior
  \item y, 320: superior/inferior
  \item z, 456: left/right
}
}
\description{
This annotation is at 25 um resolution, and is used for brain atlas annotations.
}
\details{
No parameters
}
