% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_ish.R
\name{get_exp_gene_relationships}
\alias{get_exp_gene_relationships}
\title{Query the ABA API to get relationships between ISH experiments and genes}
\usage{
get_exp_gene_relationships()
}
\value{
a data.frame with two columns: id and gene_symbol
}
\description{
No parameters
}
