% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_ish.R
\name{get_gene_aba_ish_ids}
\alias{get_gene_aba_ish_ids}
\title{Find ABA RNA ISH experiment IDs for a specific gene symbol}
\usage{
get_gene_aba_ish_ids(gene_symbol)
}
\arguments{
\item{gene_symbol}{The gene symbol to search for}
}
\value{
a character object with all of the gene symbols matching your query
}
\description{
Find ABA RNA ISH experiment IDs for a specific gene symbol
}
