% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{get_mba_ontology}
\alias{get_mba_ontology}
\title{Download the Mouse Brain Atlas ontology as a list object}
\usage{
get_mba_ontology()
}
\value{
a nested list object containing the Mouse Brain Atlas ontology
}
\description{
No parameters
}
