% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{mba_structure_id}
\alias{mba_structure_id}
\title{Get the Mouse Brain Atlas numeric structure ID(s) for given structure acronym(s)}
\usage{
mba_structure_id(acronym)
}
\arguments{
\item{acronym}{The structure acronym(s) to use.}
}
\value{
numeric value(s) for the target structure ID.
}
\description{
This retrieves the ID from an internal storage source in the package.
}
\examples{
MOp_id <- mba_structure_id("MOp")
}
