% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_ccf_meshes.R
\name{obj_to_mesh}
\alias{obj_to_mesh}
\title{Read a .obj file to get a rgl-compatible mesh object}
\usage{
obj_to_mesh(obj, material = "gray", invert_y = TRUE, yrange = c(0,
  1))
}
\arguments{
\item{obj}{A character object with the location of a .obj file.}

\item{material}{The rgl material to use for the mesh. Default is "gray". For details, see ?rgl.material.}

\item{invert_y}{Logical. Whether to flip the object in the y-dimensions. This is useful for AIBS mesh objects. Default is TRUE.}

\item{yrange}{Length 2 numeric vector indicating the range to use for inversion. Default is c(0,1).}
}
\value{
a list object with class mesh3d and shape3d.
}
\description{
Read a .obj file to get a rgl-compatible mesh object
}
