% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl_plots.R
\name{plot_brain_explorer_structures}
\alias{plot_brain_explorer_structures}
\title{Plot 3D structures of Brain explorer regions}
\usage{
plot_brain_explorer_structures(mesh_list, fg_structure,
  fg_color = "#74CAFF", fg_alpha = 1, bg_structure = NULL,
  bg_color = "#808080", bg_alpha = 0.2)
}
\arguments{
\item{mesh_list}{a named list of one or more 3D mesh objects}

\item{fg_structure}{a}
}
\description{
Plot 3D structures of Brain explorer regions
}
