% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl_plots.R
\name{plot_ccf_structure_points}
\alias{plot_ccf_structure_points}
\title{Plot a 3D CCF structure specified by its structure ID as points}
\usage{
plot_ccf_structure_points(ccf_arr, structure_id)
}
\arguments{
\item{ccf_arr}{a CCF annotation array}

\item{structure_id}{The ID number of the desired CCF structure}
}
\value{
An rgl points3d plot
}
\description{
Plot a 3D CCF structure specified by its structure ID as points
}
