% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrrd.R
\name{read_aibs_nrrd}
\alias{read_aibs_nrrd}
\title{Read AIBS .nrrd files}
\usage{
read_aibs_nrrd(file, dims = c(528, 320, 456), header_lines = 15)
}
\arguments{
\item{file}{The .nrrd file to read}

\item{dims}{The dimensions of the array. Default is for 25 nm CCF arrays: c(528,320,456)}

\item{header_lines}{The number of header lines in the .nrrd file. Default is 15.}
}
\value{
a 3 dimensional integer array with dims
}
\description{
This function reads the .nrrd format used by
AIBS, containing 32-bit integer values with 15 header lines.
}
