% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_ccf_meshes.R
\name{read_rgl_mesh_zip}
\alias{read_rgl_mesh_zip}
\title{Read a mesh object from a zip file generated by save_rgl_mesh_zip()}
\usage{
read_rgl_mesh_zip(mesh_name, zip_file, material = "gray")
}
\arguments{
\item{mesh_name}{A character object specifying the mesh name, which will be used for file names.}

\item{zip_file}{The zip file containing vectorized meshes.}

\item{material}{The rgl material to use for the mesh. Default is "gray". For details, see ?rgl.material.}
}
\value{
a list object with class mesh3d and shape3d.
}
\description{
This is currently only intended to work with triangular faces.
}
\details{
This reverses the proccessing in save_rgl_mesh_zip() to rebuild the mesh object
from binary vectors.

If you're looking for a simple way to get a brain structure, use ccf_2017_mesh, which wraps this
function, and retrieves a mesh that's included in this package.
}
\examples{
mesh_store <- system.file("extdata", "ccf_2017_meshes.zip", package = "cocoframer")
structure_id <- mba_structure_id(acronym = "MOp")
MOp_mesh <- read_rgl_mesh(structure_id,
                          zip_file = mesh_store)
}
