% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_ccf_meshes.R
\name{save_rgl_mesh_zip}
\alias{save_rgl_mesh_zip}
\title{Save a mesh object to a zip file for later use}
\usage{
save_rgl_mesh_zip(mesh, mesh_name, zip_file)
}
\arguments{
\item{mesh}{The mesh object to store.}

\item{mesh_name}{A character object specifying the mesh name, which will be used for file names.}

\item{zip_file}{The zip file to use as output.}
}
\description{
This is currently only intended to work with triangular faces.
}
\details{
This converts the vertex positions (vb) and face indexes (it) to vectors and
stores them compactly as zipped binary files.

The first value in each file is a 32-bit integer specifying the number of vertexes or positions.

No material information is retained.

Note that new data will be appended to an existing .zip file. This allows many
meshes to be stored together.
}
