% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_ish.R
\name{slice_ccf_arr}
\alias{slice_ccf_arr}
\title{Slice a 3D CCF array to get a 2D matrix}
\usage{
slice_ccf_arr(arr, slice_num, plane = "coronal")
}
\arguments{
\item{arr}{The 3D CCF array to slice}

\item{slice_num}{numeric, which slice to select.}

\item{plane}{character, which plane to use. Can be "coronal", "saggital", or "horizontal".}
}
\value{
A matrix of values from the selected slice.
}
\description{
Slice a 3D CCF array to get a 2D matrix
}
