% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{mba_structure_id}
\alias{mba_structure_id}
\title{Get the Mouse Brain Atlas numeric structure ID for a given structure acronym}
\usage{
mba_structure_id(acronym)
}
\arguments{
\item{acronym}{The structure acronym to use.}
}
\value{
a numeric value for the target structure ID.
}
\description{
This retrieves the ID from an internal storage source in the package.
}
