% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl_plots.R
\name{plot_brain_explorer_structures}
\alias{plot_brain_explorer_structures}
\title{Plot 3D structures of Brain explorer regions}
\usage{
plot_brain_explorer_structures(mesh_list, fg_structure,
  fg_color = "#74CAFF", fg_alpha = 1, bg_structure = NULL,
  bg_color = "#808080", bg_alpha = 0.2)
}
\arguments{
\item{mesh_list}{a named list of one or more 3D mesh objects}

\item{fg_structure}{The name of the structure to plot as a foreground (solid) object.}

\item{fg_color}{The color to use for the foreground object (default is light blue, "#74CAFF").}

\item{fg_alpha}{The alpha/opacity of the foreground object (default is 1).}

\item{bg_structure}{The name of the structure to plot as a background (semitransparent) object.}

\item{bg_color}{The color to use for the background object (default is medium gray, "#808080").}

\item{bg_alpha}{The alpha/opacity of the background object (default is 0.2)}
}
\value{
a 3D plot in an RGL window.
}
\description{
Plot 3D structures of Brain explorer regions
}
